src/requirements.txt
aiomysql
pymysql
fastapi[all]
scalar_fastapi
pydantic-settings
sqlmodel
alembic
dotenv